fix #21906: Properly use user set User-Agent header
This also enables the http2 plugin to be disabled without JOSM restart -- we
just set the HttpClient factory back to Http1Client::new. Ideally, users would
still restart JOSM to ensure that all classes are unloaded, but this should
make it easier to debug if an issue is due to http2 or something else.